services/diarization-worker/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pyannote.audio==3.1.1
numpy==1.26.2
huggingface-hub==0.19.4
python-multipart==0.0.6
librosa==0.10.1
pydub==0.25.1
scipy==1.11.4
matplotlib
pandas
soundfile
